erste Tests; IDs der models angepasst (long2int)

+++ app/controllers/UserController.java
@@ -18,7 +18,7 @@ import play.data.DynamicForm;
 public class UserController extends Controller {
 	
 	public static class Global {
-		public static long lUpdateId;
+		public static int lUpdateId;
 	}
 	
 	/**
@@ -39,7 +39,7 @@ public class UserController extends Controller {
 	 * @return
 	 */
 	@Transactional(readOnly=true)
-	public static Result finduser(long id) {
+	public static Result finduser(int id) {
 		Logger.info("Start");
 		User user = User.findById(id);
 		Logger.info("User searched for: " + user);
@@ -100,7 +100,7 @@ public class UserController extends Controller {
 	 * @return
 	 */
 	@Transactional
-	public static Result updateShow(long id) {
+	public static Result updateShow(int id) {
 		User user = User.findById(id);
 		Global.lUpdateId=id;
 		if (user!=null) {
@@ -147,7 +147,7 @@ public class UserController extends Controller {
 	 * @return
 	 */
 	@Transactional
-	public static Result delete(long id){
+	public static Result delete(int id){
 		User user = User.findById(id);
 		if (user==null)
 		{
